chore: unit tests code improvements
